+++ src/app/workspace/experimenting/debug_file.R
@@ -1,0 +1,2 @@
+source("GeocoderACMT.R")
+get_acs_standard_columns(year=2011, codes_of_variables_to_get=c("B08302_014", "B08302_015", "B25001_001"))